Track backup success in prometheus (#36)
When the backup script finishes, record success in prometheus. I can use
this to monitor that the scripts are running.

For now I'm using pushgateway, but maybe some type of dead man's switch
or log monitoring setup would be better in the long run. I'd like a way
of seeing logs for other services too.