Fix restic pruning policy (#38)
Previously the policy said to keep 1 backup for each daily, weekly,
monthly, and yearly. However this resulted in only 1 backup being
retained, since the most recent backup satisfied all of those criteria.

Instead what I intended was to keep 1 backup for each day (up to a
week), 1 backup for each week (up to a month), 1 backup for each month
(up to a year), and 1 backup for each year (up to 100 years). This is
the purpose of the `--keep-within-*` [policy flags], so here I switch to
those.

[policy flags]: https://restic.readthedocs.io/en/latest/060_forget.html