Increase prometheus retention (#39)
Increase from the default of 15 days to 31 days. This will make it
easier to see trends for metrics which change at a slower rate. I'm
using about 650 mb for 15 days so I can expect to use about double that
for 31 days, which is fine.

Also upgrade prometheus to the latest released version.